+++ src/verbose/values/string_value.rs
@@ -31,7 +31,7 @@ impl<'a> StringValue<'a> {
             buf.try_extend_from_slice(name.as_bytes())?;
             if buf.remaining_capacity() > 0 {
                 // Safe as capacity is checked earlier
-                unsafe { buf.push_unchecked(u8::from(0)) };
+                unsafe { buf.push_unchecked(0) };
             } else {
                 return Err(CapacityError::new(()));
             }
@@ -48,7 +48,7 @@ impl<'a> StringValue<'a> {
         buf.try_extend_from_slice(self.value.as_bytes())?;
         if buf.remaining_capacity() > 0 {
             // Safe as capacity is checked earlier
-            unsafe { buf.push_unchecked(u8::from(0)) };
+            unsafe { buf.push_unchecked(0) };
         } else {
             return Err(CapacityError::new(()));
         }
